Create LevelOne.java
This is an inherited class from the only level. created levelone  class that is used to implement the static GUI 
I've used instantiation of birds  and other objects of blocks and kept the no of them undefined in a hardcoded way as per my choice and wished to implement the pic of sample level as per my opinion.
also have won and loss button that sets to levelselectscreen with a pause button.